+++ src/gui/mod.rs
@@ -1,2 +1,3 @@
 pub mod app;
+pub mod config;
 pub mod view;
